Module runwell_ir::instr::utils

Utility types to construct or interact with some instructions.

Structs

MatchBranchInstrBuilder
Builder to construct MatchBranchInstr instructions.

MatchSelectInstrBuilder
Builder used to incrementally build up a multi-value returning match select instruction.

MatchSelectResultsIter
Iterator over the result values of all match arms of a MatchSelectInstr.