Module runwell_ir::primitive

Runwell primitive types and constant values.

Unlike Wasm the Runwell JIT supports 8-bit and 16-bit integer types. This is to generalize certain compound Wasm operations such as i32.load8_u that first load an 8-bit integer from the given address and then zero-extends it to a 32-bit integer value.

Structs

BlockEntity

A basic block entity of the Runwell IR.

F32

A f32 (32-bit floating point) value.

F64

A f64 (64-bit floating point) value.

FuncTypeEntity

A function type entity of the Runwell IR.

FunctionEntity

A function entity of the Runwell IR.

LinearMemoryEntity

A linear memory entity of the Runwell IR.

TableEntity

A table entity of the Runwell IR.

ValueEntity

An SSA value entity of the Runwell IR.

Enums

Const

A Runwell constant value.

EdgeEntity

A branching edge entity.

FloatConst

A constant floating point number value.

FloatType

Any floating point number type.

IntConst

A constant fixed-size integer value.

IntType

Any fixed-size integer type.

Type

Any Runwell supported primitive type.

Type Definitions

Block

The unique index of a basic block entity of the Runwell IR.

Edge

A unique edge entity reference.

Func

A function index.

FuncType

The unique index of a function type entity of the Runwell IR.

Mem

The unique index of a linear memory entity of the Runwell IR.

Table

The unique index of a table entity of the Runwell IR.

Value

The unique index of a value entity of the Runwell IR.