Trait runwell_ir::VisitEdgesMut

pub trait VisitEdgesMut {
    fn visit_edges_mut<V>(&mut self, visitor: V)
    where
        V: FnMut(&mut Edge) -> bool;
}

Allows to visit all edges in an instruction by mutable reference.

Implemented by all Runwell IR instructions.

Required methods

fn visit_edges_mut<V>(&mut self, visitor: V) where
V: FnMut(&mut Edge) -> bool,

Visits every edge in the instruction by mutable reference by the given visitor.

Note

The visitor returns true if it wants to continue visiting more values.

Loading content...