Trait runwell_ir::VisitValues

pub trait VisitValues {
    fn visit_values<V>(&self, visitor: V)
    where
        V: FnMut(Value) -> bool;
}

Allows to immutably visit all values in an instruction.

Implemented by all Runwell IR instructions.

Required methods

fn visit_values<V>(&self, visitor: V) where
V: FnMut(Value) -> bool,

Visits every value in the instruction by the given visitor.

Note

The visitor returns true if it wants to continue visiting more values.