Trait runwell_ir::VisitValuesMut

pub trait VisitValuesMut {
    fn visit_values_mut<V>(&mut self, visitor: V)
    where
        V: FnMut(&mut Value) -> bool;
}

Allows to visit all values in an instruction by mutable reference.

Implemented by all Runwell IR instructions.

Required methods

fn visit_values_mut<V>(&mut self, visitor: V) where
V: FnMut(&mut Value) -> bool,

Visits every value in the instruction by mutable reference by the given visitor.

Note

The visitor returns true if it wants to continue visiting more values.

Loading content...