Crate runwell_ir

Runwell IR data structures, SSA builder and IR interpreter.

Modules

instr
All Runwell IR SSA instructions.

primitive
Runwell primitive types and constant values.

Structs

ImmU32
An immediate u32 value of a Runwell IR instruction.

Indent
A single indentation.

Traits

DisplayEdge
Allows to customize display of branching edges in the Runwell IR.

DisplayInstruction
Implemented by all Runwell IR instructions to display within context.

VisitValues
Allows to immutably visit all values in an instruction.

VisitValuesMut
Allows to visit all values in an instruction by mutable reference.